fix notes
