+++ src/services/notifications/namespace.ts
@@ -1,46 +1,5 @@
 import { IAction } from 'shared/types/redux';
-import { tKeys } from 'services/i18n';
-import { SubSet } from '_helpers';
-
-export type NotificationId = string;
-export type NotificationVariant = 'info' | 'positive' | 'negative';
-export type NotificationType = Exclude<keyof typeof tKeys.services.notifications.title, 'concat'>;
-
-export type NotificationPayloadByType = SubSet<Record<NotificationType, any>, {
-  addMinter: { txHash: string },
-  addMinterFail: { txHash: string },
-  addMinterSuccess: { txHash: string },
-  createCashFlow: { txHash: string },
-  createCashFlowFail: { txHash: string },
-  createCashFlowSuccess: { txHash: string },
-  sellCashflow: { txHash: string },
-  sellCashflowSuccess: { txHash: string },
-  sellCashflowFail: { txHash: string },
-  buyCashflow: { txHash: string }
-  buyCashflowSuccess: { txHash: string },
-  buyCashflowFail: { txHash: string },
-  userPayInstallment: { txHash: string },
-  userPayInstallmentSuccess: { txHash: string },
-  userPayInstallmentFail: { txHash: string },
-}>;
-
-export const variantByType: Record<NotificationType, NotificationVariant> = {
-  addMinter: 'info',
-  addMinterFail: 'negative',
-  addMinterSuccess: 'positive',
-  createCashFlow: 'info',
-  createCashFlowFail: 'negative',
-  createCashFlowSuccess: 'positive',
-  buyCashflow: 'info',
-  buyCashflowSuccess: 'positive',
-  buyCashflowFail: 'negative',
-  sellCashflow: 'info',
-  sellCashflowSuccess: 'positive',
-  sellCashflowFail: 'negative',
-  userPayInstallment: 'info',
-  userPayInstallmentSuccess: 'positive',
-  userPayInstallmentFail: 'negative',
-};
+import { INotification, NotificationId } from 'shared/types/models';
 
 export interface IReduxState {
   data: {
@@ -49,14 +8,6 @@ export interface IReduxState {
     showingNotification: NotificationId,
   };
 }
-
-export type INotification = {
-  [key in NotificationType]: {
-    id: NotificationId;
-    type: key;
-    payload: NotificationPayloadByType[key];
-  }
-}[NotificationType];
 
 export type IPushNotification = IAction<'NOTIFICATIONS:PUSH_NOTIFICATION', INotification>;
 export type IHideNotification = IAction<'NOTIFICATIONS:HIDE_NOTIFICATION', NotificationId>;
