+++ src/shared/types/models/index.ts
@@ -2,3 +2,4 @@ export * from './common';
 export * from './dataProvider';
 export * from './signTransaction';
 export * from './cashFlow';
+export * from './notifications';
